pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "past-predictions"
version = "0.1.0"
description = "Weekly analyst target backtest pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyarrow>=12.0",
  "PyYAML>=6.0",
  "requests>=2.31",
  "typer>=0.12",
  "yfinance>=0.2.66",
  "exchange-calendars>=4.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[project.scripts]
past-predictions = "past_predictions.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
